Refactor rendering of single test response for therapist
